Added Crossplatform support

+++ TsRandomizer/PlatformHelper.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 using System.Threading;
 using Timespinner;
 using TsRandomizer.IntermediateObjects;
@@ -12,11 +14,19 @@ namespace TsRandomizer
 		{
 			var platformHelper = (PlatformHelper)Activator.CreateInstance(TimeSpinnerType.Get("Timespinner.PlatformHelper"), true);
 
-			while (Process.GetProcessesByName("Timespinner").Length == 0)
+			do {
 				Thread.Sleep(100);
 
-			foreach (var process in Process.GetProcessesByName("Timespinner"))
-				process.Kill();
+				var processes = GetTimespinnerProcesses().ToArray();
+
+				if (!processes.Any())
+					continue;
+
+				foreach (var process in processes)
+					process.Kill();
+
+				break;
+			} while (true);
 
 			return platformHelper;
 		}
@@ -24,6 +34,20 @@ namespace TsRandomizer
 		public static PlatformHelper CreateDrmFreeInstance()
 		{
 			return (PlatformHelper)Activator.CreateInstance(TimeSpinnerType.Get("Timespinner.PlatformHelper"), true);
+		}
+
+		static IEnumerable<Process> GetTimespinnerProcesses()
+		{
+			string[] processesToKill = {
+				"Timespinner",
+				"Timespinner.bin.x86",
+				"Timespinner.bin.x86_64",
+				"Timespinner.bin.osx"
+			};
+
+			foreach (var processName in processesToKill)
+				foreach (var process in Process.GetProcessesByName(processName))
+					yield return process;
 		}
 	}
 }
